+++ pyproject.toml
@@ -93,8 +93,7 @@ entry_builder = 'python_search.entry_capture.entry_inserter_gui.entry_inserter_g
 prompt_editor = 'python_search.apps.prompt_editor:main'
 entry_embeddings = 'python_search.next_item_predictor.features.entry_embeddings:main'
 entries_editor = 'python_search.entry_capture.entries_editor:main'
-register_new = 'python_search.entry_capture.register_new:main'
-register_new_launch_ui = 'python_search.entry_capture.register_new:launch_ui'
+register_new_launch_ui = 'python_search.entry_capture.entry_inserter_gui.entry_inserter_gui:launch_ui'
 share_entry = 'python_search.share:main'
 offline_evaluation = 'python_search.next_item_predictor.offline_evaluation:main'
 ps_fzf = 'python_search.search_ui.fzf:main'
