+++ python_search/apps/window_manager.py
@@ -7,10 +7,6 @@ class WindowManager:
     """
 
     @staticmethod
-    def load_from_environment() -> "WindowManager":
-        return Gnome()
-
-    @staticmethod
     def is_gnome():
         return 0 == os.system("wmctrl -m | grep -i gnome ")
 
@@ -19,45 +15,8 @@ class WindowManager:
         return 0 == os.system("wmctrl -m | grep -i xfwm4")
 
     def hide_window(self, title):
-        raise Exception("Not implemented")
+        if self.is_gnome():
+            return 0 == os.system(f"xdotool search --name '{title}' windowminimize")
 
+        raise Exception("Window manager not supported")
 
-class I3(WindowManager):
-    """
-    Contains implementation of window manager necessary functions for i3
-    The same interfaces could be implemented for other window managers
-    """
-
-    def focus_on_window_with_title(self, title) -> bool:
-        """
-        Try to focus on the window with the passed title, returns True if successful
-        """
-        cmd = f'wmctrl -a "{title}" '
-        print("Focus on window with cmd:", cmd)
-
-        result = 0 == os.system(cmd)
-
-        if result:
-            self.show_window(title)
-
-        return result
-
-    def show_window(self, title) -> bool:
-        return 0 == os.system(
-            f"unset I3SOCK ; i3-msg '[title=\"{title}\"] scratchpad show'"
-        )
-
-    def hide_window(self, title) -> bool:
-        return 0 == os.system(
-            f"sleep 0.1; i3-msg '[title=\"{title}\"]  move scratchpad'"
-        )
-
-
-class Gnome:
-    """
-    Contains implementation of window manager necessary functions for gnome
-    The same interfaces could be implemented for other window managers
-    """
-
-    def hide_window(self, title) -> bool:
-        return 0 == os.system(f"xdotool search --name '{title}' windowminimize")
