+++ python_search/entry_capture/entry_inserter_gui/entry_inserter_gui.py
@@ -12,8 +12,9 @@ from python_search.error.exception import notify_exception
 from python_search.configuration.loader import ConfigurationLoader
 from python_search.environment import is_mac
 from python_search.interpreter.interpreter_matcher import InterpreterMatcher
-from python_search.apps.notification_ui import send_notification
 from python_search.entry_type.type_detector import TypeDetector
+from python_search.entry_type.entity import infer_default_type
+from python_search.interpreter.base import BaseInterpreter
 
 
 class NewEntryGUI:
@@ -36,6 +37,43 @@ class NewEntryGUI:
         import PySimpleGUI as sg
 
         self.sg = sg
+
+
+    @notify_exception()
+    def launch_loop(self, default_type=None, default_key="", default_content=""):
+        """
+        Create a new inferred entry based on the clipboard content
+        """
+        if not default_content:
+            from python_search.apps.clipboard import Clipboard
+            default_content = Clipboard().get_content()
+
+        if not default_type:
+            default_type = infer_default_type(default_content)
+
+        from python_search.entry_capture.entry_inserter_gui.entry_inserter_gui import (
+            GuiEntryData,
+        )
+        entry_data: GuiEntryData = self.launch(
+            "New Entry",
+            default_content=default_content,
+            default_key=default_key,
+            default_type=default_type,
+        )
+        self._save_entry_data(entry_data)
+
+    def _save_entry_data(self, entry_data):
+        key = self._sanitize_key(entry_data.key)
+        interpreter: BaseInterpreter = InterpreterMatcher.build_instance(
+            self.configuration
+        ).get_interpreter_from_type(entry_data.type)
+
+        dict_entry = interpreter(entry_data.value).to_dict()
+        if entry_data.tags:
+            dict_entry["tags"] = entry_data.tags
+
+        self.entry_inserter.insert(key, dict_entry)
+
 
     @notify_exception()
     def launch(
@@ -130,6 +168,8 @@ class NewEntryGUI:
             font=(self._FONT, font_size),
             finalize=True,
         )
+
+        window.set_title("Register New")
 
         window[self._TITLE_INPUT].bind("<Escape>", "Escape")
         window[self._TITLE_INPUT].bind("<Return>", "write")
@@ -278,3 +318,5 @@ def main():
 
 if __name__ == "__main__":
     fire.Fire(NewEntryGUI().launch)
+def launch_ui():
+    fire.Fire(NewEntryGUI().launch_loop)
